requirements.txt
uetl==0.0.6
dash-core-components
dash-bootstrap-components
dash-html-components
dash-renderer
dash-table
dash-auth
dash-daq
